tools/requirements.txt
# Data visualization
matplotlib>=3.5.0

# Edit distance calculation
python-Levenshtein>=0.21.0

